Add: SAFSelectionComponent Modify: SAFHUD
Decided interface would not be enough for desired functionality. SAFISelection is to be deleted and replaced with SAFSelectionComponent.

SAFHUD went through a major update, now with custom logic for getting the actors (who have a SAFSelectionComponent) in the marquee or under the cursor. Selection is now mostly implmenented form the perspective of the SAFHUD and SAFPlayerController classes.